+++ src/rbm.jl
@@ -339,61 +339,73 @@ function gumbel_to_potts(rbm::RBM)
 end
 
 """
-    total_mean_h_from_v(rbm, v; wts = nothing)
+    total_mean_h_from_v(rbm, v; [wts])
 
 Total mean of hidden unit activations from visible activities.
 """
-function total_mean_h_from_v(rbm, v::AbstractArray; wts = nothing)
+function total_mean_h_from_v(
+        rbm, v::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.visible, v)
+    )
     inputs = inputs_h_from_v(rbm, v)
     return total_mean_from_inputs(rbm.hidden, inputs; wts)
 end
 
 """
-    total_mean_v_from_h(rbm, h; wts = nothing)
+    total_mean_v_from_h(rbm, h; [wts])
 
 Total mean of visible unit activations from given hidden activities.
 """
-function total_mean_v_from_h(rbm, h::AbstractArray; wts = nothing)
+function total_mean_v_from_h(
+        rbm, h::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.hidden, h)
+    )
     inputs = inputs_v_from_h(rbm, h)
     return total_mean_from_inputs(rbm.visible, inputs; wts)
 end
 
 """
-    total_var_h_from_v(rbm, v; wts = nothing)
+    total_var_h_from_v(rbm, v; [wts])
 
 Total variance of hidden unit activations from given visible activities.
 """
-function total_var_h_from_v(rbm, v::AbstractArray; wts = nothing)
+function total_var_h_from_v(
+        rbm, v::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.visible, v)
+    )
     inputs = inputs_h_from_v(rbm, v)
     return total_var_from_inputs(rbm.hidden, inputs; wts)
 end
 
 """
-    total_var_v_from_h(rbm, h; wts = nothing)
+    total_var_v_from_h(rbm, h; [wts])
 
 Total variance of unit activations from given hidden activities.
 """
-function total_var_v_from_h(rbm, h::AbstractArray; wts = nothing)
+function total_var_v_from_h(
+        rbm, h::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.hidden, h)
+    )
     inputs = inputs_v_from_h(rbm, h)
     return total_var_from_inputs(rbm.visible, inputs; wts)
 end
 
 """
-    total_meanvar_h_from_v(rbm, v; wts = nothing)
+    total_meanvar_h_from_v(rbm, v; [wts])
 
 Total mean and total variance of hidden unit activations from visible activities.
 """
-function total_meanvar_h_from_v(rbm, v::AbstractArray; wts = nothing)
+function total_meanvar_h_from_v(
+        rbm, v::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.visible, v)
+    )
     inputs = inputs_h_from_v(rbm, v)
     return total_meanvar_from_inputs(rbm.hidden, inputs; wts)
 end
 
 """
-    total_meanvar_v_from_h(rbm, h; wts = nothing)
+    total_meanvar_v_from_h(rbm, h; [wts])
 
 Total mean and total variance of visible unit activations from hidden activities.
 """
-function total_meanvar_v_from_h(rbm, h::AbstractArray; wts = nothing)
+function total_meanvar_v_from_h(
+        rbm, h::AbstractArray; wts::AbstractArray{<:Real} = uniform_weights(rbm.hidden, h)
+    )
     inputs = inputs_v_from_h(rbm, h)
     return total_meanvar_from_inputs(rbm.visible, inputs; wts)
 end